More app name validation code to utils.
