+++ src/components/apps/utils.js
@@ -165,3 +165,37 @@ export const getAppParameterLaunchComponent = (paramType) => {
             return Text;
     }
 };
+
+export const validateAppName = (
+    restrictedStartingChars,
+    restrictedChars,
+    value,
+    t
+) => {
+    if (!value) {
+        return t("emptyValue");
+    }
+
+    let startingCharsRegex = new RegExp("^[" + restrictedStartingChars + "]");
+    let invalid = value.match(startingCharsRegex);
+    if (invalid) {
+        return t("nameBeginsWithInvalidChars", {
+            restrictedStartingChars,
+            inValid: invalid.join(""),
+        });
+    }
+
+    // Escape each non-alphanumeric char since some are used as special chars in regex
+    let escapedRestrictedChars = restrictedChars.replace(/\W/g, "\\$&");
+    let restrictedCharsRegex = new RegExp(
+        "[" + escapedRestrictedChars + "]",
+        "g"
+    );
+    invalid = value.match(restrictedCharsRegex);
+    if (invalid) {
+        return t("nameContainsInvalidChars", {
+            restrictedChars,
+            inValid: invalid.join(""),
+        });
+    }
+};
